+++ wepay/calls/__init__.py
@@ -1,7 +1,7 @@
 from wepay.calls.oauth2 import OAuth2
 from wepay.calls.app import App
 from wepay.calls.user import User
-from wepay.calls.account import Account
+from wepay.calls.account import Account, Membership
 from wepay.calls.checkout import Checkout
 from wepay.calls.preapproval import Preapproval
 from wepay.calls.withdrawal import Withdrawal
@@ -12,9 +12,9 @@ from wepay.calls.subscription_charge import SubscriptionCharge
 from wepay.calls.batch import Batch
 
 __all__ = [
-    'OAuth2', 'App', 'User', 'Account', 'Checkout', 'Preapproval', 'Withdrawal',
-    'CreditCard', 'SubscriptionPlan', 'Subscription', 'SubscriptionCharge',
-    'Batch'
+    'OAuth2', 'App', 'User', 'Account', 'Membership', 'Checkout', 'Preapproval',
+    'Withdrawal', 'CreditCard', 'SubscriptionPlan', 'Subscription',
+    'SubscriptionCharge', 'Batch'
 ]