+++ wepay/calls/user.py
@@ -1,10 +1,15 @@
-import warnings
 from wepay.calls.base import Call
+from wepay.utils import cached_property
 
 class User(Call):
     """ The /user API calls """
 
     call_name = 'user'
+
+    @cached_property
+    def mfa(self):
+        """:class:`Membership<wepay.calls.user.MFA>` call instance"""
+        return MFA(self._api)
 
     def __call__(self, **kwargs):
         """Call documentation: `/user
@@ -28,7 +33,6 @@ class User(Call):
         return self.make_call(self, {}, kwargs)
     allowed_params = []
 
-
     def __modify(self, **kwargs):
         """Call documentation: `/user/modify
         <https://www.wepay.com/developer/reference/user#modify>`_, plus extra
@@ -51,7 +55,6 @@ class User(Call):
         return self.make_call(self.__modify, {}, kwargs)
     __modify.allowed_params = ['callback_uri']
     modify = __modify
-
 
     def __register(self, client_id, client_secret, email, scope, first_name,
                    last_name, original_ip, original_device, **kwargs):
@@ -92,7 +95,6 @@ class User(Call):
     __register.control_keywords = ['batch_mode']
     register = __register
 
-
     def __send_confirmation(self, **kwargs):
         """Call documentation: `/user/resend_confirmation
         <https://www.wepay.com/developer/reference/user#resend_confirmation>`_, plus
@@ -117,17 +119,171 @@ class User(Call):
     __send_confirmation.control_keywords = ['batch_mode']
     send_confirmation = __send_confirmation
 
-    
-    def resend_confirmation(self, *args, **kwargs):
-        """:meth:`User.send_confirmation` should be used instead.
 
-        .. warning ::
+class MFA(Call):
 
-            As of 2015-02-25 this is a deprecated call and has been renamed.
+    call_name = 'user/mfa'
+
+    def __create(self, type, **kwargs):
+        """Call documentation: `/user/mfa/create
+        <https://www.wepay.com/developer/reference/user-mfa#create>`_, plus
+        extra keyword parameter:
+        
+        :keyword bool batch_mode: turn on/off the batch_mode, see 
+           :class:`wepay.api.WePay`
+
+        :keyword str batch_reference_id: `reference_id` param for batch call,
+           see :class:`wepay.api.WePay`
+
+        :keyword str api_version: WePay API version, see
+           :class:`wepay.api.WePay`
+
         """
-        warnings.warn(
-            "/user/resend_confirmation API call has been renamed to: "
-            "/user/send_confirmation on 2015-02-25. Will be removed in 1.5 version.",
-            DeprecationWarning
-        )
-        return self.send_confirmation(*args, **kwargs)
+        params = {
+            'type': type
+        }
+        return self.make_call(self.__create, params, kwargs)
+    __create.allowed_params = [
+        'type', 'nickname', 'setup_data', 'cookie'
+    ]
+    create = __create
+
+    def __validate_cookie(self, mfa_id, cookie, **kwargs):
+        """Call documentation: `/user/mfa/validate_cookie
+        <https://www.wepay.com/developer/reference/user-mfa#validate_cookie>`_,
+        plus extra keyword parameter:
+        
+        :keyword bool batch_mode: turn on/off the batch_mode, see 
+           :class:`wepay.api.WePay`
+
+        :keyword str batch_reference_id: `reference_id` param for batch call,
+           see :class:`wepay.api.WePay`
+
+        :keyword str api_version: WePay API version, see
+           :class:`wepay.api.WePay`
+
+        """
+        params = {
+            'mfa_id': mfa_id,
+            'cookie': cookie
+        }
+        return self.make_call(self.__validate_cookie, params, kwargs)
+    __validate_cookie.allowed_params = [
+        'mfa_id', 'cookie'
+    ]
+    validate_cookie = __validate_cookie
+    
+    def __send_challenge(self, mfa_id, **kwargs):
+        """Call documentation: `/user/mfa/send_challenge
+        <https://www.wepay.com/developer/reference/user-mfa#send_challenge>`_,
+        plus extra keyword parameter:
+        
+        :keyword bool batch_mode: turn on/off the batch_mode, see 
+           :class:`wepay.api.WePay`
+
+        :keyword str batch_reference_id: `reference_id` param for batch call,
+           see :class:`wepay.api.WePay`
+
+        :keyword str api_version: WePay API version, see
+           :class:`wepay.api.WePay`
+
+        """
+        params = {
+            'mfa_id': mfa_id
+        }
+        return self.make_call(self.__send_challenge, params, kwargs)
+    __send_challenge.allowed_params = [
+        'mfa_id', 'force_voice'
+    ]
+    send_challenge = __send_challenge
+    
+    def __send_default_challenge(self, **kwargs):
+        """Call documentation: `/user/mfa/send_default_challenge
+        <https://www.wepay.com/developer/reference/user-mfa#send_default_challenge>`_,
+        plus extra keyword parameter:
+        
+        :keyword bool batch_mode: turn on/off the batch_mode, see 
+           :class:`wepay.api.WePay`
+
+        :keyword str batch_reference_id: `reference_id` param for batch call,
+           see :class:`wepay.api.WePay`
+
+        :keyword str api_version: WePay API version, see
+           :class:`wepay.api.WePay`
+
+        """
+        params = {}
+        return self.make_call(self.__send_default_challenge, params, kwargs)
+    __send_default_challenge.allowed_params = []
+    send_default_challenge = __send_default_challenge
+    
+    def __confirm(self, mfa_id, challenge, **kwargs):
+        """Call documentation: `/user/mfa/confirm
+        <https://www.wepay.com/developer/reference/user-mfa#confirm>`_,
+        plus extra keyword parameter:
+        
+        :keyword bool batch_mode: turn on/off the batch_mode, see 
+           :class:`wepay.api.WePay`
+
+        :keyword str batch_reference_id: `reference_id` param for batch call,
+           see :class:`wepay.api.WePay`
+
+        :keyword str api_version: WePay API version, see
+           :class:`wepay.api.WePay`
+
+        """
+        params = {
+            'mfa_id': mfa_id,
+            'challenge': challenge
+        }
+        return self.make_call(self.__confirm, params, kwargs)
+    __confirm.allowed_params = [
+        'mfa_id', 'challenge'
+    ]
+    confirm = __confirm
+
+    def __find(self, **kwargs):
+        """Call documentation: `/user/mfa/find
+        <https://www.wepay.com/developer/reference/user-mfa#find>`_,
+        plus extra keyword parameter:
+        
+        :keyword bool batch_mode: turn on/off the batch_mode, see 
+           :class:`wepay.api.WePay`
+
+        :keyword str batch_reference_id: `reference_id` param for batch call,
+           see :class:`wepay.api.WePay`
+
+        :keyword str api_version: WePay API version, see
+           :class:`wepay.api.WePay`
+
+        """
+        params = {}
+        return self.make_call(self.__find, params, kwargs)
+    __find.allowed_params = [
+        'challenge'
+    ]
+    find = __find
+    
+    def __modify(self, mfa_id, **kwargs):
+        """Call documentation: `/user/mfa/modify
+        <https://www.wepay.com/developer/reference/user-mfa#modify>`_,
+        plus extra keyword parameter:
+        
+        :keyword bool batch_mode: turn on/off the batch_mode, see 
+           :class:`wepay.api.WePay`
+
+        :keyword str batch_reference_id: `reference_id` param for batch call,
+           see :class:`wepay.api.WePay`
+
+        :keyword str api_version: WePay API version, see
+           :class:`wepay.api.WePay`
+
+        """
+        params = {
+            'mfa_id': mfa_id
+        }
+        return self.make_call(self.__modify, params, kwargs)
+    __modify.allowed_params = [
+        'mfa_id'
+    ]
+    modify = __modify