+++ wepay/tests/__init__.py
@@ -8,13 +8,15 @@ class CallBaseTestCase(unittest.TestCase):
         self.api = WePay(silent=False)
         self.api.call = MagicMock()
 
-    def _test_call(self, call, args, kwargs, **control_kwargs):
+    def _test_call(self, call, args, kwargs, exclude=None, **control_kwargs):
         path = call.split('/')[1:]
         fn = self.api
         for p in path:
             fn = getattr(fn, p)
         allowed_params = {x[0] for x in args}
-
+        expected_params = set(fn.allowed_params)
+        if exclude is not None:
+            expected_params = expected_params - set(exclude)
         # test required params:
         fn(*[x[1] for x in args], **control_kwargs)
         self.api.call.assert_called_once_with(
@@ -28,6 +30,6 @@ class CallBaseTestCase(unittest.TestCase):
             self.api.call.assert_called_once_with(
                 call, params=dict(args, **kwargs), **control_kwargs)
             allowed_params = allowed_params.union(kwargs)
-        self.assertEqual(allowed_params, set(fn.allowed_params))
+        self.assertEqual(allowed_params, expected_params)