+++ wepay/utils.py
@@ -25,12 +25,10 @@ class Post(object):
                 raise WePayWarning(message)
             warnings.warn(message, WePayWarning)
 
-
     def __call__(self, url, params, headers, timeout):
         if self._use_requests:
             return self._post_requests(url, params, headers, timeout)
         return self._post_urllib(url, params, headers, timeout)
-
 
     def _post_urllib(self, url, params, headers, timeout):
         data = urllib.parse.urlencode(params).encode('utf-8')
@@ -46,7 +44,6 @@ class Post(object):
         except urllib.error.URLError as exc:
             raise WePayConnectionError(exc)
         return json.loads(response.read().decode('utf-8'))
-
 
     def _post_requests(self, url, params, headers, timeout):
         data = json.dumps(params)
@@ -64,13 +61,11 @@ class Post(object):
             raise WePayConnectionError(exc)
         return response.json()
 
-
     def _raise_error(self, exc, status_code, **kwargs):
         if status_code >= 500:
             raise WePayServerError(exc, status_code, **kwargs)
         if status_code >= 400:
             raise WePayClientError(exc, status_code, **kwargs)
-
 
 
 class cached_property(object):
@@ -82,7 +77,6 @@ class cached_property(object):
         self.__module__ = fget.__module__
         self._key = "_%s" % fget.__name__
 
-
     def __get__(self, obj, cls):
         if obj is None:
             return self
@@ -91,7 +85,6 @@ class cached_property(object):
             value = self.fget(obj)
             setattr(obj, self._key, value)
         return value
-
 
     def __set__(self, obj, value):
         setattr(obj, self._key, value)